Notifications: Splits Comments and Actions Cells
